+++ src/schema.rs
@@ -33,26 +33,68 @@ pub trait SchemaType {
     fn get_type() -> crate::schema::Type;
 }
 
-/// Contains all the contract schemas for a module
+/// Contains all the contract schemas for a V0 module
 #[derive(Debug, Clone)]
-pub struct Module {
-    pub contracts: BTreeMap<String, Contract>,
+pub struct ModuleV0 {
+    pub contracts: BTreeMap<String, ContractV0>,
 }
 
-/// Describes all the schemas of a smart contract.
+/// Contains all the contract schemas for a V1 module
 #[derive(Debug, Clone)]
-pub struct Contract {
+pub struct ModuleV1 {
+    pub contracts: BTreeMap<String, ContractV1>,
+}
+
+/// Describes all the schemas of a V0 smart contract.
+/// The [Default] instance produces an empty schema.
+#[derive(Debug, Default, Clone)]
+pub struct ContractV0 {
     pub state:   Option<Type>,
     pub init:    Option<Type>,
     pub receive: BTreeMap<String, Type>,
 }
 
-impl Contract {
-    pub fn empty() -> Contract {
-        Contract {
-            state:   None,
-            init:    None,
-            receive: BTreeMap::new(),
+/// Describes all the schemas of a V1 smart contract.
+#[derive(Debug, Default, Clone)]
+/// The [Default] instance produces an empty schema.
+pub struct ContractV1 {
+    pub init:    Option<Function>,
+    pub receive: BTreeMap<String, Function>,
+}
+
+/// Describes the schema of an init or a receive function for V1 contracts.
+#[derive(Debug, Clone)]
+pub enum Function {
+    Parameter(Type),
+    ReturnValue(Type),
+    Both {
+        parameter:    Type,
+        return_value: Type,
+    },
+}
+
+impl Function {
+    /// Extract the parameter schema if it exists.
+    pub fn parameter(&self) -> Option<&Type> {
+        match self {
+            Function::Parameter(ty) => Some(ty),
+            Function::ReturnValue(_) => None,
+            Function::Both {
+                parameter,
+                ..
+            } => Some(parameter),
+        }
+    }
+
+    /// Extract the return value schema if it exists.
+    pub fn return_value(&self) -> Option<&Type> {
+        match self {
+            Function::Parameter(_) => None,
+            Function::ReturnValue(rv) => Some(rv),
+            Function::Both {
+                return_value,
+                ..
+            } => Some(return_value),
         }
     }
 }
@@ -272,24 +314,41 @@ impl Deserial for Fields {
     }
 }
 
-impl Serial for Module {
+impl Serial for ModuleV0 {
     fn serial<W: Write>(&self, out: &mut W) -> Result<(), W::Err> {
         self.contracts.serial(out)?;
         Ok(())
     }
 }
 
-impl Deserial for Module {
+impl Serial for ModuleV1 {
+    fn serial<W: Write>(&self, out: &mut W) -> Result<(), W::Err> {
+        self.contracts.serial(out)?;
+        Ok(())
+    }
+}
+
+impl Deserial for ModuleV0 {
     fn deserial<R: Read>(source: &mut R) -> ParseResult<Self> {
         let len: u32 = source.get()?;
         let contracts = deserial_map_no_length_no_order_check(source, len as usize)?;
-        Ok(Module {
+        Ok(ModuleV0 {
             contracts,
         })
     }
 }
 
-impl Serial for Contract {
+impl Deserial for ModuleV1 {
+    fn deserial<R: Read>(source: &mut R) -> ParseResult<Self> {
+        let len: u32 = source.get()?;
+        let contracts = deserial_map_no_length_no_order_check(source, len as usize)?;
+        Ok(ModuleV1 {
+            contracts,
+        })
+    }
+}
+
+impl Serial for ContractV0 {
     fn serial<W: Write>(&self, out: &mut W) -> Result<(), W::Err> {
         self.state.serial(out)?;
         self.init.serial(out)?;
@@ -298,17 +357,75 @@ impl Serial for Contract {
     }
 }
 
-impl Deserial for Contract {
+impl Serial for ContractV1 {
+    fn serial<W: Write>(&self, out: &mut W) -> Result<(), W::Err> {
+        self.init.serial(out)?;
+        self.receive.serial(out)?;
+        Ok(())
+    }
+}
+
+impl Deserial for ContractV0 {
     fn deserial<R: Read>(source: &mut R) -> ParseResult<Self> {
         let state = source.get()?;
         let init = source.get()?;
         let len: u32 = source.get()?;
         let receive = deserial_map_no_length_no_order_check(source, len as usize)?;
-        Ok(Contract {
+        Ok(ContractV0 {
             state,
             init,
             receive,
         })
+    }
+}
+
+impl Deserial for ContractV1 {
+    fn deserial<R: Read>(source: &mut R) -> ParseResult<Self> {
+        let init = source.get()?;
+        let len: u32 = source.get()?;
+        let receive = deserial_map_no_length_no_order_check(source, len as usize)?;
+        Ok(ContractV1 {
+            init,
+            receive,
+        })
+    }
+}
+
+impl Serial for Function {
+    fn serial<W: Write>(&self, out: &mut W) -> Result<(), W::Err> {
+        match self {
+            Function::Parameter(parameter) => {
+                out.write_u8(0)?;
+                parameter.serial(out)
+            }
+            Function::ReturnValue(return_value) => {
+                out.write_u8(1)?;
+                return_value.serial(out)
+            }
+            Function::Both {
+                parameter,
+                return_value,
+            } => {
+                out.write_u8(2)?;
+                parameter.serial(out)?;
+                return_value.serial(out)
+            }
+        }
+    }
+}
+
+impl Deserial for Function {
+    fn deserial<R: Read>(source: &mut R) -> ParseResult<Self> {
+        let idx = source.read_u8()?;
+        match idx {
+            0 => Ok(Function::Parameter(source.get()?)),
+            1 => Ok(Function::ReturnValue(source.get()?)),
+            2 => Ok(Function::Both {
+                parameter:    source.get()?,
+                return_value: source.get()?,
+            }),
+            _ => Err(ParseError::default()),
+        }
     }
 }